Move test_callNativesOnNewClassLoaders to a native test
Move test_callNativesOnNewClassLoaders to a separated test bucket which
supports native tests, otherwise other tests like JCL_TEST_Java-Lang
will fail due to the incompatible LIBPATH set by
ADD_JVM_LIB_DIR_TO_LIBPATH.
